refactor(cost): delete the orphaned parallel scan path
Commit 62d67db moved production onto CostSummaryBuilder.makeScan and the
budgeted, resumable scanSessions(session:) path. That orphaned the older
whole-corpus scan one commit before f4ed571 added parallelism to it, so
CostScanParallel has never run in production — it existed only to be
tested.

Deletes the dead cluster: CostScanParallel, CostSummaryBuilder.makeSummary
and scanSources, both ClaudeCostScanner.scanSessions(since:) overloads,
CodexCostScanner.scanSessions(since:) and its two scanRollouts(directory:)
seams, and CostScanFileSystem.transcriptFiles.

Codex fixture tests still want a fixed, readable file order and a plain
end-to-end read of each rollout, which the budgeted path (newest-mtime
first, through a byte-offset cache) does not give. They now drive
CodexCostScanner.parseRollout(at:windows:) through a test-target helper,
CostScanFixtureScan, which reproduces the old sorted-path serial walk.

The width-1-vs-width-8 comparisons and the parseInOrder unit tests go with
the code they covered. The behaviour they also happened to pin —
deduplication, session counting, and surviving unreadable or malformed
transcripts — moves to CostScanCorpusTests, running against the surviving
entry points.

No production behaviour change: nothing deleted here was reachable from
the app.

Verification: swift test (1922 tests, 0 failures), swiftlint clean.
