+++ MeterBar/Services/CodexCostScanner.swift
@@ -26,19 +26,6 @@ enum CodexCostScanner {
         return result
     }()
 
-    nonisolated static func scanSessions(since cutoffDate: Date) -> ScanWindows<CodexScanContext> {
-        let codexDir = URL(fileURLWithPath: CodexHomeDirectory.path(), isDirectory: true)
-        let logsDatabase = codexDir.appendingPathComponent("logs_2.sqlite")
-        var windows = Self.scanWindows(cutoff: cutoffDate)
-
-        for directory in Self.rolloutDirectories(in: codexDir) {
-            Self.scanRollouts(directory: directory, windows: &windows)
-        }
-        Self.scanSQLiteLogs(database: logsDatabase, since: cutoffDate, windows: &windows)
-
-        return windows
-    }
-
     /// Seeds both windows the way the two separate scans used to seed themselves:
     /// `earliestDate` starts at now and only decreases, `latestDate` starts at the
     /// window floor (the cutoff for the period, `.distantPast` for lifetime) and
@@ -147,6 +134,9 @@ enum CodexCostScanner {
     nonisolated private static let turnContextMarker = Data("\"turn_context\"".utf8)
     nonisolated private static let sessionMetaMarker = Data("\"session_meta\"".utf8)
 
+    /// Turns one rollout into the usage events it contributes, in the order the
+    /// scan applies them.
+    ///
     /// Internal (not private) so the rollout parsing — the Codex counterpart to
     /// `parseSessionFile`, and where CLI-vs-app cost divergence hides — can be
     /// fixture-tested against a temp directory.
@@ -154,44 +144,8 @@ enum CodexCostScanner {
     /// A `token_count` event names neither the model nor the front end, so the
     /// scan streams each file in order and carries the last `turn_context`
     /// model and the opening `session_meta` originator forward. Attribution is
-    /// per file: state resets on every rollout.
-    ///
-    /// Unlike Claude's per-file dedup, `CodexScanContext.eventKeys` spans every
-    /// rollout *and* the SQLite log, first event wins. So only the parse fans
-    /// out: each file is turned into an ordered list of usage events on a
-    /// worker, and those lists are applied to the shared windows serially, in
-    /// sorted-path order. Merging per-file contexts instead would have to
-    /// resolve dedup collisions after the fact, and the winner would depend on
-    /// which worker finished first.
-    nonisolated static func scanRollouts(
-        directory: URL,
-        windows: inout ScanWindows<CodexScanContext>,
-        width: Int = CostScanParallel.defaultWidth
-    ) {
-        // The per-file mtime prefilter is gone on purpose: the lifetime window
-        // needs every file, and the period window is still bounded by the
-        // per-event timestamp check inside `ScanWindows.update` (an event is
-        // never newer than the file holding it).
-        let files = CostScanFileSystem.transcriptFiles(in: directory, options: [.skipsHiddenFiles])
-
-        CostScanParallel.parseInOrder(
-            files,
-            width: width,
-            parse: { Self.parseRollout(at: $0) },
-            fold: { events in
-                for event in events {
-                    Self.apply(event, windows: &windows)
-                }
-            }
-        )
-    }
-
-    /// Turns one rollout into the usage events it contributes, in the order the
-    /// serial scan would have applied them.
-    ///
-    /// Everything this reads is scoped to the single file — attribution carried
-    /// across rollouts would label one session's spend with another's model —
-    /// which is precisely why it is safe to run on a worker thread.
+    /// per file: state resets on every rollout, since attribution carried
+    /// across rollouts would label one session's spend with another's model.
     nonisolated static func parseRollout(at fileURL: URL) -> [CodexUsageEvent] {
         var events: [CodexUsageEvent] = []
         var rollout = CodexRolloutContext()
@@ -230,32 +184,20 @@ enum CodexCostScanner {
 
     /// Streams one rollout end to end into `windows`.
     ///
-    /// Attribution state is created here and dies here: carried across rollouts
-    /// it would label one session's spend with another's model.
-    nonisolated private static func parseRollout(
+    /// Serial by construction. Unlike Claude's per-file dedup,
+    /// `CodexScanContext.eventKeys` spans every rollout *and* the SQLite log,
+    /// first event wins — so the order files reach here is the order that
+    /// decides the answer.
+    ///
+    /// Internal (not private) so a fixture scan can drive a whole directory
+    /// through it.
+    nonisolated static func parseRollout(
         at fileURL: URL,
         windows: inout ScanWindows<CodexScanContext>
     ) {
         for event in Self.parseRollout(at: fileURL) {
             Self.apply(event, windows: &windows)
         }
-    }
-
-    /// Single-window entry point kept for callers that only care about one
-    /// period. Scans into `context` as the period window and discards lifetime.
-    nonisolated static func scanRollouts(
-        directory: URL,
-        since cutoffDate: Date,
-        context: inout CodexScanContext,
-        width: Int = CostScanParallel.defaultWidth
-    ) {
-        var windows = ScanWindows(
-            period: context,
-            lifetime: CodexScanContext(earliestDate: Date(), latestDate: .distantPast),
-            cutoff: cutoffDate
-        )
-        Self.scanRollouts(directory: directory, windows: &windows, width: width)
-        context = windows.period
     }
 
     // MARK: - Budgeted, resumable scan
