+++ MeterBarTests/CostTrackerTests.swift
@@ -312,7 +312,11 @@ final class CostTrackerTests: XCTestCase {
 
         let account = ClaudeCodeAccount(id: UUID(), name: "demo", configDirectory: configDir.path)
         let cutoff = FlexibleISO8601.date(from: "2026-06-01T00:00:00Z")!
-        let windows = ClaudeCostScanner.scanSessions(since: cutoff, claudeAccounts: [account])
+        let session = CostScanSession(cutoff: cutoff, options: .unlimited)
+        let windows = ClaudeCostScanner.scanRoots(
+            ClaudeCostScanner.projectRoots(accounts: [account]),
+            session: session
+        )
 
         XCTAssertEqual(windows.period.projects["www/demo/project"]?.input, 10)
     }
@@ -320,7 +324,7 @@ final class CostTrackerTests: XCTestCase {
     // MARK: - Codex rollout scan
 
     /// Writes a `.jsonl` into an `archived_sessions` directory and returns that
-    /// directory (the argument `CodexCostScanner.scanRollouts` expects). Every file is
+    /// directory (the argument `CostScanFixtureScan.codexRollouts` expects). Every file is
     /// read regardless of its modification date, so the lifetime window sees
     /// pre-cutoff lines too.
     private func writeCodexArchive(lines: [String]) throws -> URL {
@@ -415,7 +419,7 @@ final class CostTrackerTests: XCTestCase {
         let cutoff = FlexibleISO8601.date(from: "2026-01-01T00:00:00Z")!
         var context = makeContext(cutoff: cutoff)
 
-        CodexCostScanner.scanRollouts(directory: dir, since: cutoff, context: &context)
+        CostScanFixtureScan.codexRollouts(in: dir, since: cutoff, context: &context)
 
         XCTAssertEqual(context.totals.input, 1_000)
         // output accumulates the reasoning tokens on the daily/model rollups, but
@@ -432,7 +436,7 @@ final class CostTrackerTests: XCTestCase {
         let cutoff = FlexibleISO8601.date(from: "2026-01-01T00:00:00Z")!
         var context = makeContext(cutoff: cutoff)
 
-        CodexCostScanner.scanRollouts(directory: dir, since: cutoff, context: &context)
+        CostScanFixtureScan.codexRollouts(in: dir, since: cutoff, context: &context)
 
         // Identical events collapse to one via the dedup key.
         XCTAssertEqual(context.totals.input, 1_000)
@@ -447,7 +451,7 @@ final class CostTrackerTests: XCTestCase {
         let cutoff = FlexibleISO8601.date(from: "2026-01-01T00:00:00Z")!
         var context = makeContext(cutoff: cutoff)
 
-        CodexCostScanner.scanRollouts(directory: dir, since: cutoff, context: &context)
+        CostScanFixtureScan.codexRollouts(in: dir, since: cutoff, context: &context)
 
         // Only the post-cutoff line is counted.
         XCTAssertEqual(context.totals.input, 100)
@@ -466,7 +470,7 @@ final class CostTrackerTests: XCTestCase {
         let cutoff = FlexibleISO8601.date(from: "2026-01-01T00:00:00Z")!
         var context = makeContext(cutoff: cutoff)
 
-        CodexCostScanner.scanRollouts(directory: dir, since: cutoff, context: &context)
+        CostScanFixtureScan.codexRollouts(in: dir, since: cutoff, context: &context)
 
         XCTAssertEqual(context.totals.input, 1_000)
         XCTAssertEqual(context.totals.output, 500)
@@ -488,7 +492,7 @@ final class CostTrackerTests: XCTestCase {
         let cutoff = FlexibleISO8601.date(from: "2026-01-01T00:00:00Z")!
         var context = makeContext(cutoff: cutoff)
 
-        CodexCostScanner.scanRollouts(directory: dir, since: cutoff, context: &context)
+        CostScanFixtureScan.codexRollouts(in: dir, since: cutoff, context: &context)
 
         XCTAssertEqual(context.totals.input, 42)
         XCTAssertEqual(context.totals.output, 7)
@@ -508,7 +512,7 @@ final class CostTrackerTests: XCTestCase {
         let cutoff = FlexibleISO8601.date(from: "2026-01-01T00:00:00Z")!
         var context = makeContext(cutoff: cutoff)
 
-        CodexCostScanner.scanRollouts(directory: dir, since: cutoff, context: &context)
+        CostScanFixtureScan.codexRollouts(in: dir, since: cutoff, context: &context)
 
         XCTAssertEqual(Array(context.modelTotals.keys), ["gpt-5.6-sol"])
         XCTAssertEqual(Array(context.originTotals.keys), ["Codex Desktop"])
@@ -537,7 +541,7 @@ final class CostTrackerTests: XCTestCase {
         let cutoff = FlexibleISO8601.date(from: "2026-01-01T00:00:00Z")!
         var context = makeContext(cutoff: cutoff)
 
-        CodexCostScanner.scanRollouts(directory: dir, since: cutoff, context: &context)
+        CostScanFixtureScan.codexRollouts(in: dir, since: cutoff, context: &context)
 
         XCTAssertEqual(context.projectTotals["www/genfeed/apps/admin"]?.input, 1_000)
         XCTAssertNil(context.projectTotals[CostProjectAttribution.unknownProjectID])
@@ -551,7 +555,7 @@ final class CostTrackerTests: XCTestCase {
         let cutoff = FlexibleISO8601.date(from: "2026-01-01T00:00:00Z")!
         var context = makeContext(cutoff: cutoff)
 
-        CodexCostScanner.scanRollouts(directory: dir, since: cutoff, context: &context)
+        CostScanFixtureScan.codexRollouts(in: dir, since: cutoff, context: &context)
 
         XCTAssertEqual(context.projectTotals[CostProjectAttribution.unknownProjectID]?.input, 1_000)
     }
@@ -567,7 +571,7 @@ final class CostTrackerTests: XCTestCase {
         let cutoff = FlexibleISO8601.date(from: "2026-01-01T00:00:00Z")!
         var context = makeContext(cutoff: cutoff)
 
-        CodexCostScanner.scanRollouts(directory: dir, since: cutoff, context: &context)
+        CostScanFixtureScan.codexRollouts(in: dir, since: cutoff, context: &context)
 
         XCTAssertEqual(context.projectModelTotals["www/genfeed/apps/admin"]?["gpt-5.6-sol"]?.input, 1_000)
 
@@ -607,7 +611,7 @@ final class CostTrackerTests: XCTestCase {
         let cutoff = FlexibleISO8601.date(from: "2026-01-01T00:00:00Z")!
         var context = makeContext(cutoff: cutoff)
 
-        CodexCostScanner.scanRollouts(directory: dir, since: cutoff, context: &context)
+        CostScanFixtureScan.codexRollouts(in: dir, since: cutoff, context: &context)
 
         XCTAssertEqual(context.modelTotals["gpt-5.6-sol"]?.input, 100)
         XCTAssertEqual(context.modelTotals["gpt-5.6-luna"]?.input, 7)
@@ -622,7 +626,7 @@ final class CostTrackerTests: XCTestCase {
         let cutoff = FlexibleISO8601.date(from: "2026-01-01T00:00:00Z")!
         var context = makeContext(cutoff: cutoff)
 
-        CodexCostScanner.scanRollouts(directory: dir, since: cutoff, context: &context)
+        CostScanFixtureScan.codexRollouts(in: dir, since: cutoff, context: &context)
 
         XCTAssertEqual(Array(context.modelTotals.keys), ["Unknown model"])
         XCTAssertEqual(Array(context.originTotals.keys), ["Codex CLI"])
@@ -638,7 +642,7 @@ final class CostTrackerTests: XCTestCase {
         let cutoff = FlexibleISO8601.date(from: "2026-01-01T00:00:00Z")!
         var context = makeContext(cutoff: cutoff)
 
-        CodexCostScanner.scanRollouts(directory: dir, since: cutoff, context: &context)
+        CostScanFixtureScan.codexRollouts(in: dir, since: cutoff, context: &context)
 
         XCTAssertEqual(Array(context.modelTotals.keys), ["gpt-5.5"])
     }
@@ -659,7 +663,7 @@ final class CostTrackerTests: XCTestCase {
         let cutoff = FlexibleISO8601.date(from: "2026-01-01T00:00:00Z")!
         var context = makeContext(cutoff: cutoff)
 
-        CodexCostScanner.scanRollouts(directory: dir, since: cutoff, context: &context)
+        CostScanFixtureScan.codexRollouts(in: dir, since: cutoff, context: &context)
 
         XCTAssertEqual(context.modelTotals["gpt-5.6-sol"]?.input, 100)
         XCTAssertEqual(context.modelTotals["Unknown model"]?.input, 7)
@@ -681,7 +685,7 @@ final class CostTrackerTests: XCTestCase {
         let cutoff = FlexibleISO8601.date(from: "2026-01-01T00:00:00Z")!
         var context = makeContext(cutoff: cutoff)
 
-        CodexCostScanner.scanRollouts(directory: dir, since: cutoff, context: &context)
+        CostScanFixtureScan.codexRollouts(in: dir, since: cutoff, context: &context)
 
         XCTAssertEqual(context.modelTotals["gpt-5.6-sol"]?.input, 107)
         XCTAssertNil(context.modelTotals["Unknown model"], "the file names the model — nothing is unknown")
@@ -701,7 +705,7 @@ final class CostTrackerTests: XCTestCase {
         let cutoff = FlexibleISO8601.date(from: "2026-01-01T00:00:00Z")!
         var context = makeContext(cutoff: cutoff)
 
-        CodexCostScanner.scanRollouts(directory: dir, since: cutoff, context: &context)
+        CostScanFixtureScan.codexRollouts(in: dir, since: cutoff, context: &context)
 
         XCTAssertEqual(context.modelTotals["gpt-5.6-sol"]?.input, 107)
         XCTAssertEqual(context.modelTotals["gpt-5.6-luna"]?.input, 3)
@@ -722,7 +726,7 @@ final class CostTrackerTests: XCTestCase {
         let cutoff = FlexibleISO8601.date(from: "2026-01-01T00:00:00Z")!
         var context = makeContext(cutoff: cutoff)
 
-        CodexCostScanner.scanRollouts(directory: dir, since: cutoff, context: &context)
+        CostScanFixtureScan.codexRollouts(in: dir, since: cutoff, context: &context)
 
         XCTAssertEqual(context.modelTotals["Unknown model"]?.input, 100)
         XCTAssertEqual(context.modelTotals["gpt-5.6-sol"]?.input, 7)
@@ -739,7 +743,7 @@ final class CostTrackerTests: XCTestCase {
         let cutoff = FlexibleISO8601.date(from: "2026-01-01T00:00:00Z")!
         var context = makeContext(cutoff: cutoff)
 
-        CodexCostScanner.scanRollouts(directory: dir, since: cutoff, context: &context)
+        CostScanFixtureScan.codexRollouts(in: dir, since: cutoff, context: &context)
 
         let firstDay = Calendar.current.startOfDay(
             for: try XCTUnwrap(FlexibleISO8601.date(from: "2026-06-14T12:00:00Z"))
@@ -774,7 +778,7 @@ final class CostTrackerTests: XCTestCase {
         let cutoff = FlexibleISO8601.date(from: "2026-01-01T00:00:00Z")!
         var context = makeContext(cutoff: cutoff)
 
-        CodexCostScanner.scanRollouts(directory: dir, since: cutoff, context: &context)
+        CostScanFixtureScan.codexRollouts(in: dir, since: cutoff, context: &context)
 
         XCTAssertEqual(context.modelTotals["gpt-5.6-sol"]?.input, 100)
         XCTAssertEqual(context.sessionIDs, ["conv-x"])
@@ -796,7 +800,7 @@ final class CostTrackerTests: XCTestCase {
         var context = makeContext(cutoff: cutoff)
 
         for directory in CodexCostScanner.rolloutDirectories(in: root) {
-            CodexCostScanner.scanRollouts(directory: directory, since: cutoff, context: &context)
+            CostScanFixtureScan.codexRollouts(in: directory, since: cutoff, context: &context)
         }
 
         XCTAssertEqual(context.totals.input, 100)
@@ -892,7 +896,7 @@ final class CostTrackerTests: XCTestCase {
         let cutoff = FlexibleISO8601.date(from: "2026-01-01T00:00:00Z")!
         var windows = CodexCostScanner.scanWindows(cutoff: cutoff)
 
-        CodexCostScanner.scanRollouts(directory: dir, windows: &windows)
+        CostScanFixtureScan.codexRollouts(in: dir, windows: &windows)
 
         XCTAssertEqual(windows.period.totals.input, 100)
         XCTAssertEqual(windows.period.sessionIDs, ["new"])
@@ -906,7 +910,7 @@ final class CostTrackerTests: XCTestCase {
         let cutoff = FlexibleISO8601.date(from: "2026-01-01T00:00:00Z")!
         var windows = CodexCostScanner.scanWindows(cutoff: cutoff)
 
-        CodexCostScanner.scanRollouts(directory: dir, windows: &windows)
+        CostScanFixtureScan.codexRollouts(in: dir, windows: &windows)
 
         XCTAssertEqual(windows.period.totals.input, 1_000)
         XCTAssertEqual(windows.lifetime.totals.input, 1_000)
